Playlist previous button working.
Seems that Musicovery has died so now we use LastFM to get recommendations.

When you enter on the website, artist and track recommendations are automatically loaded based on the local storage content or, if there is no data stored, we load recommendations decided by us.

Artist Recommendations section can be loaded now.